backend/requirements.txt
python-dotenv
fastapi
uvicorn
python-socketio
faster-whisper
google-generativeai
deep-translator
gTTS
python-multipart
pydantic
aiofiles
soundfile
librosa
numpy
torch
ffmpeg-python
